refactor: 상태관리container, UI 파일 분리
